requirements.txt
openai
python-dotenv
langdetect
google-auth
google-auth-oauthlib
google-api-python-client
packaging
deep-translator
flask